third_party/blink/web_tests/external/wpt/editing/run/bold_3001-last-expected.txt

This is a testharness.js-based test.
[FAIL] [["stylewithcss","false"],["bold",""]] "<span style=font-weight:100>fo[o</span><span style=font-weight:200>b]ar</span>" queryCommandState("stylewithcss") after
  assert_equals: Wrong result returned expected false but got true
[FAIL] [["stylewithcss","false"],["bold",""]] "<span style=font-weight:100>fo[o</span><span style=font-weight:200>b]ar</span>" queryCommandValue("bold") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["stylewithcss","false"],["bold",""]] "<span style=font-weight:100>fo[o</span><span style=font-weight:200>b]ar</span>" queryCommandState("bold") after
  assert_equals: Wrong result returned expected true but got false
[FAIL] [["stylewithcss","false"],["bold",""]] "<span style=font-weight:100>fo[o</span><span style=font-weight:200>b]ar</span>" queryCommandValue("bold") after
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["stylewithcss","false"],["bold",""]] "abc<b>[d<span contenteditable=\\"false\\"><b>e</b></span>f]</b>ghi" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "abcd<span contenteditable=\\"false\\"><b>e</b></span>fghi" but got "abcd<span contenteditable=\\"false\\" style=\\"\\"><b>e</b></span>fghi"
[FAIL] [["stylewithcss","false"],["bold",""]] "abc<b>[d<span contenteditable=\\"false\\"><span contenteditable><b>e</b></span></span>f]</b>ghi" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "abcd<span contenteditable=\\"false\\"><span contenteditable=\\"\\">e</span></span>fghi" but got "abcd<span contenteditable=\\"false\\" style=\\"\\"><span contenteditable=\\"\\">e</span></span>fghi"
[FAIL] [["stylewithcss","false"],["bold",""]] "abc[<i>def</i>]ghi" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "abc<b><i>def</i></b>ghi" but got "abc<i><b>def</b></i>ghi"
[FAIL] [["stylewithcss","false"],["bold",""]] "abc{<i>def</i>}ghi" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "abc<b><i>def</i></b>ghi" but got "abc<i><b>def</b></i>ghi"
Harness: the test ran to completion.
